# core/docs/swagger.yaml
basePath: /
definitions:
  domain.AlertRule:
    properties:
      duration:
        example: 2m
        type: string
      enabled:
        type: boolean
      id:
        type: string
      metric_name:
        type: string
      name:
        type: string
      operator:
        type: string
      recovery_policy:
        type: string
      severity:
        type: string
      target_id:
        type: string
      threshold:
        type: number
    type: object
  domain.Event:
    properties:
      container_name:
        type: string
      event_type:
        type: string
      id:
        type: integer
      message:
        type: string
      node_id:
        type: string
      payload:
        additionalProperties: {}
        type: object
      severity:
        type: string
      source:
        type: string
      target_id:
        type: string
      timestamp:
        type: string
    type: object
  domain.Incident:
    properties:
      description:
        type: string
      id:
        type: integer
      node_id:
        type: string
      resolved_at:
        type: string
      rule_id:
        type: string
      severity:
        type: string
      started_at:
        type: string
      status:
        type: string
      target_id:
        type: string
      value:
        type: number
    type: object
  domain.MetricPoint:
    properties:
      container_name:
        type: string
      metric_name:
        type: string
      node_id:
        type: string
      target_id:
        type: string
      timestamp:
        type: string
      unit:
        type: string
      value:
        type: number
    type: object
  domain.MetricSnapshot:
    properties:
      block_read_bytes:
        type: integer
      block_write_bytes:
        type: integer
      container_name:
        type: string
      cpu_usage_percent:
        type: number
      memory_usage_bytes:
        type: integer
      network_rx_bytes:
        type: integer
      network_tx_bytes:
        type: integer
      node_id:
        type: string
      target_id:
        type: string
      timestamp:
        type: string
    type: object
  domain.RecoveryAction:
    properties:
      action_type:
        type: string
      finished_at:
        type: string
      id:
        type: integer
      incident_id:
        type: integer
      result_message:
        type: string
      started_at:
        type: string
      status:
        type: string
      target_id:
        type: string
    type: object
  domain.Target:
    properties:
      created_at:
        type: string
      external_id:
        type: string
      id:
        type: string
      labels:
        additionalProperties: {}
        type: object
      last_seen_at:
        type: string
      name:
        type: string
      node_id:
        type: string
      source:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  http.retryRecoveryResponse:
    properties:
      id:
        example: "1"
        type: string
      status:
        example: retry accepted
        type: string
    type: object
  http.statusResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
info:
  contact: {}
  description: MVP API for Docker container monitoring, alerting, incidents, and recovery
    actions.
  title: Container Monitoring API
  version: 0.1.0
paths:
  /api/v1/alert-rules:
    get:
      description: Returns enabled and disabled threshold alert rules.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AlertRule'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: List alert rules
      tags:
      - alert-rules
    post:
      consumes:
      - application/json
      description: Creates a threshold alert rule. Missing id, operator, duration,
        severity, recovery_policy, and enabled fields are defaulted by the API.
      parameters:
      - description: Alert rule
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/domain.AlertRule'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.AlertRule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Create alert rule
      tags:
      - alert-rules
  /api/v1/alert-rules/{id}:
    delete:
      description: Deletes a threshold alert rule.
      parameters:
      - description: Alert rule ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Delete alert rule
      tags:
      - alert-rules
    patch:
      consumes:
      - application/json
      description: Updates a threshold alert rule.
      parameters:
      - description: Alert rule ID
        in: path
        name: id
        required: true
        type: string
      - description: Alert rule
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/domain.AlertRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AlertRule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Update alert rule
      tags:
      - alert-rules
  /api/v1/events:
    get:
      description: Returns recent container events from PostgreSQL.
      parameters:
      - description: Maximum rows
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: List events
      tags:
      - events
  /api/v1/incidents:
    get:
      description: Returns recent incidents ordered by start time.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Incident'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: List incidents
      tags:
      - incidents
  /api/v1/incidents/{id}:
    get:
      description: Returns one incident by ID.
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Incident'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Get incident
      tags:
      - incidents
  /api/v1/incidents/{id}/ack:
    post:
      description: Marks an incident as acknowledged.
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Acknowledge incident
      tags:
      - incidents
  /api/v1/incidents/{id}/resolve:
    post:
      description: Marks an incident as resolved and sets resolved_at.
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Resolve incident
      tags:
      - incidents
  /api/v1/metrics/history:
    get:
      description: Returns metric history points from ClickHouse.
      parameters:
      - description: Target ID
        in: query
        name: target_id
        type: string
      - description: Metric name
        in: query
        name: metric_name
        type: string
      - description: RFC3339 start time
        in: query
        name: from
        type: string
      - description: RFC3339 end time
        in: query
        name: to
        type: string
      - description: Maximum rows
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.MetricPoint'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Metric history
      tags:
      - metrics
  /api/v1/metrics/latest:
    get:
      description: Returns recent metric snapshots from ClickHouse.
      parameters:
      - description: Target ID
        in: query
        name: target_id
        type: string
      - description: Maximum rows
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.MetricSnapshot'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Latest metrics
      tags:
      - metrics
  /api/v1/recovery-actions:
    get:
      description: Returns recent recovery action attempts and results.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.RecoveryAction'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: List recovery actions
      tags:
      - recovery
  /api/v1/recovery-actions/{id}/retry:
    post:
      description: Accepts a retry request for a recovery action placeholder endpoint.
      parameters:
      - description: Recovery action ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.retryRecoveryResponse'
      summary: Retry recovery action
      tags:
      - recovery
  /api/v1/targets:
    get:
      description: Returns discovered Docker container targets.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Target'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: List targets
      tags:
      - targets
    post:
      consumes:
      - application/json
      description: Creates or registers a monitored container target.
      parameters:
      - description: Target
        in: body
        name: target
        required: true
        schema:
          $ref: '#/definitions/domain.Target'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Target'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Create target
      tags:
      - targets
  /api/v1/targets/{id}:
    delete:
      description: Deletes a target record from the configuration store.
      parameters:
      - description: Target ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Delete target
      tags:
      - targets
    get:
      description: Returns one discovered Docker container target.
      parameters:
      - description: Target ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Target'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Get target
      tags:
      - targets
    patch:
      consumes:
      - application/json
      description: Updates target metadata used by the admin panel.
      parameters:
      - description: Target ID
        in: path
        name: id
        required: true
        type: string
      - description: Target
        in: body
        name: target
        required: true
        schema:
          $ref: '#/definitions/domain.Target'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Target'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Update target
      tags:
      - targets
  /api/v1/targets/{id}/events:
    get:
      description: Documents target-scoped event access.
      parameters:
      - description: Target ID
        in: path
        name: id
        required: true
        type: string
      - description: Maximum rows
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Target events pointer
      tags:
      - events
  /api/v1/targets/{id}/metrics:
    get:
      description: Documents target-scoped metric history access.
      parameters:
      - description: Target ID
        in: path
        name: id
        required: true
        type: string
      - description: Metric name
        in: query
        name: metric_name
        type: string
      - description: Maximum rows
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.MetricPoint'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.statusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Target metrics pointer
      tags:
      - metrics
  /health:
    get:
      description: Reports that the HTTP process is alive.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Health check
      tags:
      - health
  /ready:
    get:
      description: Reports that the HTTP process is ready to serve requests.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.statusResponse'
      summary: Readiness check
      tags:
      - health
schemes:
- http
swagger: "2.0"